Add macOS target
Uses DirectHW and pciutils to do hardware accesses.
Supports features that you only see in the DOS or EFI targets.
Supports all versions of macOS but 10.3 and earlier are untested and may require additional work.
Supports PPC and Intel. Arm is untested and may require updates to DirectHW that may or may not be possible.
Any kext that is controlling the PCI card should probably be unloaded while using lsiutil.